requirements.txt
django>=2.1
djangorestframework==3.10.3
psycopg2